requirements.txt
tornado==6.0.4
diffusers
numpy==1.21.6
openpyxl==3.0.10
PyOpenGL==3.1.6
python-dateutil==2.8.2
python-slugify==6.1.2
python-utils==3.4.5
Pillow==7.1.2
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
torch==1.12.1
torchaudio==0.12.1
torchsummary==1.5.1
torchtext==0.13.1
torchvision==0.13.1
zipp==3.10.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
transformers
accelerate